Moved bitfields derive code into separate crate
