+++ c2rust-bitfields-derive/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "c2rust-bitfields-derive"
+version = "0.1.0"
+authors = [
+    "The C2Rust Project Developers <c2rust@immunant.com>",
+    "Daniel Kolsoi <djk@immunant.com>",
+]
+license = "BSD-3-Clause"
+homepage = "https://c2rust.com/"
+repository = "https://github.com/immunant/c2rust/tree/master/c2rust-bitfields-derive"
+edition = "2018"
+description = "C-compatible struct bitfield derive implementation used in the C2Rust project"
+readme = "README.md"
+
+[dependencies]
+syn = { version = "0.15", features = ["full"] }
+quote = "0.6"
+
+[lib]
+proc-macro = true
